Multiple inheritance and overriding example
Write an abstract class Employee with three variables name, sal and Grosssal , suitable
constructors, print method and two abstract methods calculategrosssalary() and
annualincrement(). Create a manager sub-class of employee with hr as a member
variable and write an implementation of the abstract method. Also, create a subclass of
manager as sales manager with a commission as a member variable and override the
calulategrosssalary() method.